Allow to re-subscribe to a topic
There is a bug in rosgo where once a subscriber for a topic is shutdown, no new
subscribers for the same topic would function.  The problem is that shutting
down a subscriber will unregister the subscriber from the master, but wouldn't
update node.subscribers map keeping track of active subscribers. So next
subscribers to the same topic only add a callback to a now defunkt subscriber.

This CL passes a callback to the subscriber to remove itself from the node on
the shutdown.  I chose to use a callback instead of using channels, because
spinning up another goroutine for subscriber management seemed like an overkill
for this particular task.